backend/requirements.txt
python-jose==3.3.0
alembic==1.13.1
fastapi==0.110.3
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
SQLAlchemy==2.0.29
uvicorn==0.29.0
psycopg2-binary==2.9.9
envparse==0.2.0
httpx==0.11.1
databases==0.9.0
asyncpg==0.29.0